+++ lib/src/components/form/form.dart
@@ -42,7 +42,6 @@ enum FormValidationMode {
   initial,
   changed,
   submitted,
-  waiting,
 }
 
 class ValidationMode<T> extends Validator<T> {
@@ -326,6 +325,13 @@ class LengthValidator extends Validator<String> {
   FutureOr<ValidationResult?> validate(
       BuildContext context, String? value, FormValidationMode state) {
     if (value == null) {
+      if (min != null) {
+        return InvalidResult(
+            message ??
+                Localizations.of(context, ShadcnLocalizations)
+                    .formLengthLessThan(min!),
+            state: state);
+      }
       return null;
     }
     ShadcnLocalizations localizations =
@@ -896,18 +902,51 @@ class CompositeValidator<T> extends Validator<T> {
 abstract class ValidationResult {
   final FormValidationMode state;
   const ValidationResult({required this.state});
+  FormKey get key;
+  ValidationResult attach(FormKey key);
 }
 
 class ReplaceResult<T> extends ValidationResult {
   final T value;
+  final FormKey? _key;
 
-  const ReplaceResult(this.value, {required super.state});
+  const ReplaceResult(this.value, {required super.state}) : _key = null;
+
+  const ReplaceResult.attached(this.value,
+      {required FormKey key, required super.state})
+      : _key = key;
+
+  @override
+  FormKey get key {
+    assert(_key != null, 'The result has not been attached to a key');
+    return _key!;
+  }
+
+  @override
+  ReplaceResult<T> attach(FormKey key) {
+    return ReplaceResult.attached(value, key: key, state: state);
+  }
 }
 
 class InvalidResult extends ValidationResult {
   final String message;
+  final FormKey? _key;
 
-  const InvalidResult(this.message, {required super.state});
+  const InvalidResult(this.message, {required super.state}) : _key = null;
+  const InvalidResult.attached(this.message,
+      {required FormKey key, required super.state})
+      : _key = key;
+
+  @override
+  FormKey get key {
+    assert(_key != null, 'The result has not been attached to a key');
+    return _key!;
+  }
+
+  @override
+  InvalidResult attach(FormKey key) {
+    return InvalidResult.attached(message, key: key, state: state);
+  }
 }
 
 class FormValidityNotification extends Notification {
@@ -926,6 +965,14 @@ class FormKey<T> extends LocalKey {
 
   bool isInstanceOf(dynamic value) {
     return value is T;
+  }
+
+  T? getValue(FormMapValues values) {
+    return values.getValue(this);
+  }
+
+  T? operator [](FormMapValues values) {
+    return values.getValue(this);
   }
 
   @override
@@ -982,14 +1029,20 @@ class FormEntry<T> extends StatefulWidget {
 }
 
 mixin FormFieldHandle {
-  FutureOr<ValidationResult?> reportNewFormValue(Object? value);
+  FutureOr<ValidationResult?> reportNewFormValue<T>(T? value);
   FutureOr<ValidationResult?> revalidate();
   ValueListenable<ValidationResult?>? get validity;
 }
 
+class _FormEntryCachedValue {
+  Object? value;
+
+  _FormEntryCachedValue(this.value);
+}
+
 class FormEntryState extends State<FormEntry> with FormFieldHandle {
   FormController? _controller;
-  Object? _cachedValue;
+  _FormEntryCachedValue? _cachedValue;
   final ValueNotifier<ValidationResult?> _validity = ValueNotifier(null);
 
   @override
@@ -1050,24 +1103,17 @@ class FormEntryState extends State<FormEntry> with FormFieldHandle {
   }
 
   @override
-  FutureOr<ValidationResult?> reportNewFormValue(Object? value) {
-    bool isSameType = widget.key.isInstanceOf(value);
-    assert(isSameType, () {
-      throw FlutterError.fromParts([
-        ErrorSummary('Invalid value type'),
-        ErrorDescription(
-            'Expecting ${widget.key.type} but received ${value.runtimeType}'),
-        ErrorHint('Make sure the value type matches the key type'),
-        ErrorHint('Key: FormKey<${widget.key.type}>'),
-      ]);
-    });
+  FutureOr<ValidationResult?> reportNewFormValue<T>(T? value) {
+    bool isSameType = widget.key.type == T;
     if (!isSameType) {
+      var parentLookup = Data.maybeFind<FormFieldHandle>(context);
+      return parentLookup?.reportNewFormValue<T>(value);
+    }
+    var cachedValue = _cachedValue;
+    if (cachedValue != null && cachedValue.value == value) {
       return null;
     }
-    if (_cachedValue == value) {
-      return null;
-    }
-    _cachedValue = value;
+    _cachedValue = _FormEntryCachedValue(value);
     return _controller?.attach(context, widget.key, value, widget.validator);
   }
 
@@ -1122,8 +1168,8 @@ class _FormEntryHandleInterceptor with FormFieldHandle {
   const _FormEntryHandleInterceptor(this.handle, this.onValueReported);
 
   @override
-  FutureOr<ValidationResult?> reportNewFormValue(Object? value) {
-    return handle?.reportNewFormValue(value);
+  FutureOr<ValidationResult?> reportNewFormValue<T>(T? value) {
+    return handle?.reportNewFormValue<T>(value);
   }
 
   @override
@@ -1172,8 +1218,22 @@ class FormValueState<T> {
   int get hashCode => Object.hash(value, validator);
 }
 
+typedef FormMapValues = Map<FormKey, dynamic>;
+
 typedef FormSubmitCallback = void Function(
-    BuildContext context, Map<FormKey, dynamic> values);
+    BuildContext context, FormMapValues values);
+
+extension FormMapValuesExtension on FormMapValues {
+  T? getValue<T>(FormKey<T> key) {
+    Object? value = this[key];
+    if (value == null) {
+      return null;
+    }
+    assert(key.isInstanceOf(value),
+        'The value for key $key is not of type ${key.type}');
+    return value as T?;
+  }
+}
 
 class Form extends StatefulWidget {
   final FormController? controller;
@@ -1185,9 +1245,16 @@ class Form extends StatefulWidget {
   State<Form> createState() => FormState();
 }
 
+class _ValidatorResultStash {
+  final FutureOr<ValidationResult?> result;
+  final FormValidationMode state;
+
+  const _ValidatorResultStash(this.result, this.state);
+}
+
 class FormController extends ChangeNotifier {
   final Map<FormKey, FormValueState> _attachedInputs = {};
-  final Map<FormKey, FutureOr<ValidationResult?>> _validity = {};
+  final Map<FormKey, _ValidatorResultStash> _validity = {};
 
   bool _disposed = false;
 
@@ -1204,11 +1271,35 @@ class FormController extends ChangeNotifier {
   }
 
   Map<FormKey, FutureOr<ValidationResult?>> get validities {
-    return Map.unmodifiable(_validity);
+    return Map.unmodifiable(_validity.map((key, value) {
+      return MapEntry(key, value.result);
+    }));
+  }
+
+  Map<FormKey, ValidationResult> get errors {
+    final errors = <FormKey, ValidationResult>{};
+    for (var entry in _validity.entries) {
+      var result = entry.value.result;
+      if (result is Future<ValidationResult?>) {
+        errors[entry.key] =
+            WaitingResult.attached(state: entry.value.state, key: entry.key);
+      } else if (result != null) {
+        errors[entry.key] = result;
+      }
+    }
+    return errors;
   }
 
   FutureOr<ValidationResult?>? getError(FormKey key) {
-    return _validity[key];
+    return _validity[key]?.result;
+  }
+
+  ValidationResult? getSyncError(FormKey key) {
+    var result = _validity[key]?.result;
+    if (result is Future<ValidationResult?>) {
+      return WaitingResult.attached(state: _validity[key]!.state, key: key);
+    }
+    return result;
   }
 
   FormValueState? getState(FormKey key) {
@@ -1222,12 +1313,13 @@ class FormController extends ChangeNotifier {
       var value = entry.value;
       if (value.validator != null) {
         var future = value.validator!.validate(context, value.value, state);
-        if (_validity[key] != future) {
-          _validity[key] = future;
+        if (_validity[key]?.result != future) {
           if (future is Future<ValidationResult?>) {
+            _validity[key] = _ValidatorResultStash(future, state);
             future.then((value) {
-              if (_validity[key] == future) {
-                _validity[key] = value;
+              if (_validity[key]?.result == future) {
+                _validity[key] =
+                    _ValidatorResultStash(value?.attach(key), state);
                 WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
                   if (_disposed) {
                     return;
@@ -1236,6 +1328,8 @@ class FormController extends ChangeNotifier {
                 });
               }
             });
+          } else {
+            _validity[key] = _ValidatorResultStash(future, state);
           }
           changed = true;
         }
@@ -1265,12 +1359,12 @@ class FormController extends ChangeNotifier {
     _attachedInputs[key] = state;
     // validate
     var future = validator?.validate(context, value, lifecycle);
-    _validity[key] = future;
     if (future is Future<ValidationResult?>) {
+      _validity[key] = _ValidatorResultStash(future, lifecycle);
       future.then((value) {
         // resolve the future and store synchronous value
-        if (_validity[key] == future) {
-          _validity[key] = value;
+        if (_validity[key]?.result == future) {
+          _validity[key] = _ValidatorResultStash(value?.attach(key), lifecycle);
           WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
             if (_disposed) {
               return;
@@ -1279,6 +1373,8 @@ class FormController extends ChangeNotifier {
           });
         }
       });
+    } else {
+      _validity[key] = _ValidatorResultStash(future, lifecycle);
     }
     // check for revalidation
     Map<FormKey, FutureOr<ValidationResult?>> revalidate = {};
@@ -1301,11 +1397,11 @@ class FormController extends ChangeNotifier {
       attachedInput = FormValueState(
           value: attachedInput.value, validator: attachedInput.validator);
       _attachedInputs[k] = attachedInput;
-      _validity[k] = future;
       if (future is Future<ValidationResult?>) {
+        _validity[k] = _ValidatorResultStash(future, lifecycle);
         future.then((value) {
-          if (_validity[k] == future) {
-            _validity[k] = value;
+          if (_validity[k]?.result == future) {
+            _validity[k] = _ValidatorResultStash(value?.attach(k), lifecycle);
             WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
               if (_disposed) {
                 return;
@@ -1314,6 +1410,8 @@ class FormController extends ChangeNotifier {
             });
           }
         });
+      } else {
+        _validity[k] = _ValidatorResultStash(future, lifecycle);
       }
     }
     WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
@@ -1400,7 +1498,20 @@ class FormEntryErrorBuilder extends StatelessWidget {
 }
 
 class WaitingResult extends ValidationResult {
-  const WaitingResult({required super.state});
+  final FormKey? _key;
+  const WaitingResult.attached({required FormKey key, required super.state})
+      : _key = key;
+
+  @override
+  FormKey get key {
+    assert(_key != null, 'The result has not been attached to a key');
+    return _key!;
+  }
+
+  @override
+  WaitingResult attach(FormKey key) {
+    return WaitingResult.attached(key: key, state: state);
+  }
 }
 
 class FormErrorBuilder extends StatelessWidget {
@@ -1414,46 +1525,11 @@ class FormErrorBuilder extends StatelessWidget {
   Widget build(BuildContext context) {
     final formController = Data.maybeOf<FormController>(context);
     if (formController != null) {
-      return AnimatedBuilder(
-        animation: formController,
+      return ListenableBuilder(
+        listenable: formController,
         child: child,
         builder: (context, child) {
-          final errors = formController.validities;
-          // future builder
-          return FutureBuilder<List<MapEntry<FormKey, ValidationResult?>>>(
-            future: Future.wait(errors.entries.map((entry) {
-              var key = entry.key;
-              var value = entry.value;
-              if (value is Future<ValidationResult?>) {
-                return value.then((value) => MapEntry(key, value));
-              }
-              return Future.value(MapEntry(key, value));
-            })),
-            builder: (context, snapshot) {
-              if (snapshot.connectionState == ConnectionState.waiting) {
-                return builder(
-                    context,
-                    {
-                      for (var entry in formController._attachedInputs.entries)
-                        entry.key: const WaitingResult(
-                            state: FormValidationMode.waiting)
-                    },
-                    child);
-              }
-              if (snapshot.hasData) {
-                final errors = <FormKey, ValidationResult>{};
-                for (var entry in snapshot.data!) {
-                  var key = entry.key;
-                  var value = entry.value;
-                  if (value != null) {
-                    errors[key] = value;
-                  }
-                }
-                return builder(context, errors, child);
-              }
-              return builder(context, {}, child);
-            },
-          );
+          return builder(context, formController.errors, child);
         },
       );
     }
@@ -1541,12 +1617,12 @@ extension FormExtension on BuildContext {
   FutureOr<SubmissionResult> _chainedSubmitForm(
       Map<FormKey, Object?> values,
       Map<FormKey, ValidationResult> errors,
-      Iterator<MapEntry<FormKey, FutureOr<ValidationResult?>>> iterator) {
+      Iterator<MapEntry<FormKey, _ValidatorResultStash>> iterator) {
     if (!iterator.moveNext()) {
       return SubmissionResult(values, errors);
     }
     var entry = iterator.current;
-    var value = entry.value;
+    var value = entry.value.result;
     if (value is Future<ValidationResult?>) {
       return value.then((value) {
         if (value != null) {
@@ -1592,7 +1668,7 @@ mixin FormValueSupplier<T, X extends StatefulWidget> on State<X> {
       return;
     }
     final currentCounter = ++_futureCounter;
-    var validationResult = state.reportNewFormValue(value);
+    var validationResult = state.reportNewFormValue<T>(value);
     if (validationResult is Future<ValidationResult?>) {
       validationResult.then((value) {
         if (_futureCounter == currentCounter) {
@@ -1864,7 +1940,7 @@ class FormTableLayout extends StatelessWidget {
   }
 }
 
-class SubmitButton extends StatefulWidget {
+class SubmitButton extends StatelessWidget {
   final AbstractButtonStyle? style;
   final Widget child;
   final Widget? loading;
@@ -1903,122 +1979,60 @@ class SubmitButton extends StatefulWidget {
   });
 
   @override
-  widgets.State<SubmitButton> createState() => _SubmitButtonState();
-}
-
-class _SubmitButtonState extends widgets.State<SubmitButton> {
-  FutureOr<bool>? _future;
-  FormController? _controller;
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-    var oldController = _controller;
-    var newController = Data.maybeOf<FormController>(context);
-    if (oldController != newController) {
-      oldController?.removeListener(_onControllerChanged);
-      newController?.addListener(_onControllerChanged);
-      _controller = newController;
-      if (_controller != null) {
-        _onControllerChanged();
-      }
-    }
-  }
-
-  void _onControllerChanged() {
-    if (!mounted) {
-      return;
-    }
-    setState(() {
-      _future = _hasError();
-    });
-  }
-
-  int count = 0;
-
-  FutureOr<bool> _hasError() {
-    if (_controller == null) {
-      return false;
-    }
-    return _chainedHasError(_controller!._validity.entries.iterator);
-  }
-
-  FutureOr<bool> _chainedHasError(
-      Iterator<MapEntry<FormKey, FutureOr<ValidationResult?>>> iterator) {
-    if (!iterator.moveNext()) {
-      return false;
-    }
-    var entry = iterator.current;
-    var value = entry.value;
-    if (value is Future<ValidationResult?>) {
-      return value.then((value) {
-        if (value != null) {
-          return true;
-        }
-        return _chainedHasError(iterator);
-      });
-    }
-    if (value != null) {
-      return true;
-    }
-    return _chainedHasError(iterator);
-  }
-
-  @override
-  Widget build(widgets.BuildContext context) {
-    var hasError = _future;
-    if (hasError is Future) {
-      // loading
-      return Button(
-        leading: widget.loadingLeading ?? widget.leading,
-        trailing: widget.loadingTrailing ?? widget.trailing,
-        alignment: widget.alignment,
-        disableHoverEffect: widget.disableHoverEffect,
-        enabled: false,
-        enableFeedback: false,
-        disableTransition: widget.disableTransition,
-        focusNode: widget.focusNode,
-        style: widget.style ?? const ButtonStyle.primary(),
-        child: widget.loading ?? widget.child,
-      );
-    }
-    if (hasError == true) {
-      return Button(
-        leading: widget.errorLeading ?? widget.leading,
-        trailing: widget.errorTrailing ?? widget.trailing,
-        alignment: widget.alignment,
-        disableHoverEffect: widget.disableHoverEffect,
-        enabled: false,
-        enableFeedback: true,
-        disableTransition: widget.disableTransition,
-        focusNode: widget.focusNode,
-        style: widget.style ?? const ButtonStyle.primary(),
-        child: widget.error ?? widget.child,
-      );
-    }
-    return Button(
-      trailing: widget.trailing,
-      leading: widget.leading,
-      alignment: widget.alignment,
-      disableHoverEffect: widget.disableHoverEffect,
-      enabled: widget.enabled,
-      enableFeedback: widget.enableFeedback,
-      onPressed: () {
-        setState(() {
-          var submissionResult = context.submitForm();
-          if (submissionResult is Future<SubmissionResult>) {
-            _future = submissionResult.then((value) {
-              return value.errors.isNotEmpty;
-            });
-          } else {
-            _future = submissionResult.errors.isNotEmpty;
-          }
+  widgets.Widget build(widgets.BuildContext context) {
+    return FormErrorBuilder(
+      builder: (context, errors, child) {
+        var hasWaitingError = errors.values.any((element) {
+          return element is WaitingResult;
         });
+        var hasError = errors.values.any((element) {
+          return element is InvalidResult;
+        });
+        if (hasWaitingError) {
+          return Button(
+            leading: loadingLeading ?? leading,
+            trailing: loadingTrailing ?? trailing,
+            alignment: alignment,
+            disableHoverEffect: disableHoverEffect,
+            enabled: false,
+            enableFeedback: false,
+            disableTransition: disableTransition,
+            focusNode: focusNode,
+            style: style ?? const ButtonStyle.primary(),
+            child: loading ?? child!,
+          );
+        }
+        if (hasError) {
+          return Button(
+            leading: errorLeading ?? leading,
+            trailing: errorTrailing ?? trailing,
+            alignment: alignment,
+            disableHoverEffect: disableHoverEffect,
+            enabled: false,
+            enableFeedback: true,
+            disableTransition: disableTransition,
+            focusNode: focusNode,
+            style: style ?? const ButtonStyle.primary(),
+            child: error ?? child!,
+          );
+        }
+        return Button(
+          leading: leading,
+          trailing: trailing,
+          alignment: alignment,
+          disableHoverEffect: disableHoverEffect,
+          enabled: enabled ?? true,
+          enableFeedback: enableFeedback ?? true,
+          onPressed: () {
+            context.submitForm();
+          },
+          disableTransition: disableTransition,
+          focusNode: focusNode,
+          style: style ?? const ButtonStyle.primary(),
+          child: child!,
+        );
       },
-      disableTransition: widget.disableTransition,
-      focusNode: widget.focusNode,
-      style: widget.style ?? const ButtonStyle.primary(),
-      child: widget.child,
+      child: child,
     );
   }
 }