+++ lib/src/components/form/formatter.dart
@@ -1,14 +1,12 @@
 import 'dart:math';
 
+import 'package:expressions/expressions.dart';
 import 'package:flutter/services.dart';
 
-
-TextSelection contraintToNewText(
-    TextEditingValue oldValue, TextEditingValue newValue) {
+TextSelection contraintToNewText(TextEditingValue newValue, String newText) {
   return TextSelection(
-    baseOffset: newValue.selection.baseOffset.clamp(0, newValue.text.length),
-    extentOffset:
-        newValue.selection.extentOffset.clamp(0, newValue.text.length),
+    baseOffset: newValue.selection.baseOffset.clamp(0, newText.length),
+    extentOffset: newValue.selection.extentOffset.clamp(0, newText.length),
   );
 }
 
@@ -21,6 +19,16 @@ class TextInputFormatters {
 
   static TextInputFormatter integerOnly({int? min, int? max}) {
     return _IntegerOnlyFormatter(min: min, max: max);
+  }
+
+  static TextInputFormatter digitsOnly(
+      {double? min, double? max, int? decimalDigits}) {
+    return _DoubleOnlyFormatter(
+        min: min, max: max, decimalDigits: decimalDigits);
+  }
+
+  static TextInputFormatter mathExpression({Map<String, dynamic>? context}) {
+    return _MathExpressionFormatter(context: context);
   }
 
   const TextInputFormatters._();
@@ -76,20 +84,142 @@ class _IntegerOnlyFormatter extends TextInputFormatter {
     TextEditingValue oldValue,
     TextEditingValue newValue,
   ) {
-    int? value = int.tryParse(newValue.text);
+    String newText = newValue.text;
+    if (newText.isEmpty) {
+      return newValue;
+    }
+    bool negate = newText.startsWith('-');
+    if (negate) {
+      newText = newText.substring(1);
+    }
+    int? value = int.tryParse(newText);
     if (value == null) {
+      if (negate) {
+        return const TextEditingValue(
+          text: '-',
+          selection: TextSelection.collapsed(offset: 1),
+        );
+      }
       return oldValue;
     }
-    if (min != null && value < min!) {
+    if (min != null && value <= min!) {
       value = min!;
     }
-    if (max != null && value > max!) {
+    if (max != null && value >= max!) {
       value = max!;
     }
-    var newText = value.toString();
+    newText = value.toString();
+    if (negate) {
+      newText = '-$newText';
+    }
     return TextEditingValue(
       text: newText,
-      selection: contraintToNewText(oldValue, newValue),
+      selection: contraintToNewText(newValue, newText),
+    );
+  }
+}
+
+class _DoubleOnlyFormatter extends TextInputFormatter {
+  final double? min;
+  final double? max;
+  final int? decimalDigits;
+
+  _DoubleOnlyFormatter({this.min, this.max, this.decimalDigits});
+
+  @override
+  TextEditingValue formatEditUpdate(
+    TextEditingValue oldValue,
+    TextEditingValue newValue,
+  ) {
+    String newText = newValue.text;
+    if (newText.isEmpty) {
+      return newValue;
+    }
+    bool negate = newText.startsWith('-');
+    if (negate) {
+      newText = newText.substring(1);
+    }
+    bool endsWithDot = newText.endsWith('.');
+    if (endsWithDot) {
+      newText = newText.substring(0, newText.length - 1);
+    }
+    double? value = double.tryParse(newText);
+    if (value == null) {
+      if (negate) {
+        return const TextEditingValue(
+          text: '-',
+          selection: TextSelection.collapsed(offset: 1),
+        );
+      }
+      return oldValue;
+    }
+    if (min != null && value <= min!) {
+      value = min!;
+      endsWithDot = false;
+    }
+    if (max != null && value >= max!) {
+      value = max!;
+      endsWithDot = false;
+    }
+    // var newText = value.toString();
+    if (decimalDigits != null) {
+      newText = value.toStringAsFixed(decimalDigits!);
+    } else {
+      newText = value.toString();
+    }
+    if (newText.contains('.')) {
+      while (newText.endsWith('0')) {
+        newText = newText.substring(0, newText.length - 1);
+      }
+      if (newText.endsWith('.')) {
+        newText = newText.substring(0, newText.length - 1);
+      }
+    }
+    if (endsWithDot) {
+      newText += '.';
+    }
+    if (negate) {
+      newText = '-$newText';
+    }
+    return TextEditingValue(
+      text: newText,
+      selection: contraintToNewText(newValue, newText),
+    );
+  }
+}
+
+class _MathExpressionFormatter extends TextInputFormatter {
+  final Map<String, dynamic>? context;
+  const _MathExpressionFormatter({this.context});
+  @override
+  TextEditingValue formatEditUpdate(
+    TextEditingValue oldValue,
+    TextEditingValue newValue,
+  ) {
+    String newText = newValue.text;
+    Object? result;
+    try {
+      Expression expression = Expression.parse(newText);
+      ExpressionEvaluator evaluator = const ExpressionEvaluator();
+      result = evaluator.eval(expression, context ?? {});
+      if (result is! num) {
+        result = '';
+      }
+    } catch (e) {
+      result = '';
+    }
+    String resultText = result.toString();
+    if (resultText.contains('.')) {
+      while (resultText.endsWith('0')) {
+        resultText = resultText.substring(0, resultText.length - 1);
+      }
+      if (resultText.endsWith('.')) {
+        resultText = resultText.substring(0, resultText.length - 1);
+      }
+    }
+    return TextEditingValue(
+      text: resultText,
+      selection: contraintToNewText(newValue, resultText),
     );
   }
 }