+++ lib/src/components/form/text_field.dart
@@ -16,6 +16,9 @@ import 'package:shadcn_flutter/src/components/layout/hidden.dart';
 
 import '../../../shadcn_flutter.dart';
 
+import 'package:flutter/material.dart' as material;
+import 'package:flutter/cupertino.dart' as cupertino;
+
 export 'package:flutter/services.dart'
     show
         SmartDashesType,
@@ -523,9 +526,77 @@ mixin TextInput on Widget {
   SpellCheckConfiguration? get spellCheckConfiguration;
   UndoHistoryController? get undoController;
   List<InputFeature> get features;
+  List<TextInputFormatter>? get submitFormatters;
+  bool get skipInputFeatureFocusTraversal;
 }
 
 class TextField extends StatefulWidget with TextInput {
+  static EditableTextContextMenuBuilder nativeContextMenuBuilder() {
+    return (context, editableTextState) {
+      return material.AdaptiveTextSelectionToolbar.editableText(
+          editableTextState: editableTextState);
+    };
+  }
+
+  static EditableTextContextMenuBuilder cupertinoContextMenuBuilder() {
+    return (context, editableTextState) {
+      return cupertino.CupertinoAdaptiveTextSelectionToolbar.editableText(
+          editableTextState: editableTextState);
+    };
+  }
+
+  static EditableTextContextMenuBuilder materialContextMenuBuilder() {
+    return (context, editableTextState) {
+      final anchors = editableTextState.contextMenuAnchors;
+      return material.TextSelectionToolbar(
+        anchorAbove: anchors.primaryAnchor,
+        anchorBelow: anchors.secondaryAnchor == null
+            ? anchors.primaryAnchor
+            : anchors.secondaryAnchor!,
+        children: _getMaterialButtons(
+            context, editableTextState.contextMenuButtonItems),
+      );
+    };
+  }
+
+  static List<Widget> _getMaterialButtons(
+    BuildContext context,
+    List<ContextMenuButtonItem> buttonItems,
+  ) {
+    final List<Widget> buttons = <Widget>[];
+    for (int i = 0; i < buttonItems.length; i++) {
+      final ContextMenuButtonItem buttonItem = buttonItems[i];
+      buttons.add(
+        material.TextSelectionToolbarTextButton(
+          padding: material.TextSelectionToolbarTextButton.getPadding(
+              i, buttonItems.length),
+          onPressed: buttonItem.onPressed,
+          alignment: AlignmentDirectional.centerStart,
+          child: Text(_getMaterialButtonLabel(context, buttonItem)),
+        ),
+      );
+    }
+    return buttons;
+  }
+
+  static String _getMaterialButtonLabel(
+      BuildContext context, ContextMenuButtonItem buttonItem) {
+    final localizations = material.MaterialLocalizations.of(context);
+    return switch (buttonItem.type) {
+      ContextMenuButtonType.cut => localizations.cutButtonLabel,
+      ContextMenuButtonType.copy => localizations.copyButtonLabel,
+      ContextMenuButtonType.paste => localizations.pasteButtonLabel,
+      ContextMenuButtonType.selectAll => localizations.selectAllButtonLabel,
+      ContextMenuButtonType.delete =>
+        localizations.deleteButtonTooltip.toUpperCase(),
+      ContextMenuButtonType.lookUp => localizations.lookUpButtonLabel,
+      ContextMenuButtonType.searchWeb => localizations.searchWebButtonLabel,
+      ContextMenuButtonType.share => localizations.shareButtonLabel,
+      ContextMenuButtonType.liveTextInput => localizations.scanTextButtonLabel,
+      ContextMenuButtonType.custom => '',
+    };
+  }
+
   const TextField({
     super.key,
     this.groupId = EditableText,
@@ -600,6 +671,8 @@ class TextField extends StatefulWidget with TextInput {
     this.filled = false,
     this.statesController,
     this.features = const [],
+    this.submitFormatters = const [],
+    this.skipInputFeatureFocusTraversal = true,
   })  : assert(obscuringCharacter.length == 1),
         smartDashesType = smartDashesType ??
             (obscureText ? SmartDashesType.disabled : SmartDashesType.enabled),
@@ -629,6 +702,9 @@ class TextField extends StatefulWidget with TextInput {
             (maxLines == 1 ? TextInputType.text : TextInputType.multiline),
         enableInteractiveSelection =
             enableInteractiveSelection ?? (!readOnly || !obscureText);
+
+  @override
+  final bool skipInputFeatureFocusTraversal;
 
   @override
   final List<InputFeature> features;
@@ -837,6 +913,9 @@ class TextField extends StatefulWidget with TextInput {
 
   @override
   final WidgetStatesController? statesController;
+
+  @override
+  final List<TextInputFormatter>? submitFormatters;
 
   static Widget _defaultContextMenuBuilder(
     BuildContext context,
@@ -1096,6 +1175,8 @@ class TextField extends StatefulWidget with TextInput {
     TextMagnifierConfiguration? magnifierConfiguration,
     SpellCheckConfiguration? spellCheckConfiguration,
     List<InputFeature>? features,
+    List<TextInputFormatter>? submitFormatters,
+    bool? skipInputFeatureFocusTraversal,
   }) {
     return TextField(
       key: key ?? this.key,
@@ -1176,6 +1257,9 @@ class TextField extends StatefulWidget with TextInput {
       spellCheckConfiguration:
           spellCheckConfiguration ?? this.spellCheckConfiguration,
       features: features ?? this.features,
+      submitFormatters: submitFormatters ?? this.submitFormatters,
+      skipInputFeatureFocusTraversal:
+          skipInputFeatureFocusTraversal ?? this.skipInputFeatureFocusTraversal,
     );
   }
 }
@@ -1246,11 +1330,14 @@ class TextFieldState extends State<TextField>
       _createLocalController(widget.initialValue != null
           ? TextEditingValue(text: widget.initialValue!)
           : null);
+    } else {
+      widget.controller!.addListener(_handleControllerChanged);
     }
     _effectiveFocusNode.canRequestFocus = widget.enabled;
     _effectiveFocusNode.addListener(_handleFocusChanged);
     _statesController = widget.statesController ?? WidgetStatesController();
-    formValue = widget.controller?.text ?? widget.initialValue ?? '';
+    String effectiveText = widget.controller?.text ?? widget.initialValue ?? '';
+    formValue = effectiveText.isEmpty ? null : effectiveText;
     for (final feature in widget.features) {
       final state = feature.createState();
       state._attached = _AttachedInputFeature(feature, state);
@@ -1347,6 +1434,8 @@ class TextFieldState extends State<TextField>
   void _handleControllerChanged() {
     _effectiveText.value = effectiveController.text;
     _effectiveSelection.value = effectiveController.selection;
+    formValue =
+        effectiveController.text.isEmpty ? null : effectiveController.text;
   }
 
   void _createLocalController([TextEditingValue? value]) {
@@ -1389,6 +1478,22 @@ class TextFieldState extends State<TextField>
       // highlight.
     });
     _statesController.update(WidgetState.focused, _effectiveFocusNode.hasFocus);
+    if (!_effectiveFocusNode.hasFocus) {
+      _formatSubmit();
+    }
+  }
+
+  void _formatSubmit() {
+    if (widget.submitFormatters != null) {
+      TextEditingValue value = effectiveController.value;
+      for (var formatter in widget.submitFormatters!) {
+        value = formatter.formatEditUpdate(value, value);
+      }
+      if (value != effectiveController.value) {
+        effectiveController.value = value;
+        widget.onChanged?.call(value.text);
+      }
+    }
   }
 
   bool _shouldShowSelectionHandles(SelectionChangedCause? cause) {
@@ -1465,7 +1570,8 @@ class TextFieldState extends State<TextField>
   bool get _hasDecoration {
     return widget.placeholder != null ||
         widget.leading != null ||
-        widget.trailing != null;
+        widget.trailing != null ||
+        widget.features.isNotEmpty;
   }
 
   // Provide default behavior if widget.textAlignVertical is not set.
@@ -1539,10 +1645,20 @@ class TextFieldState extends State<TextField>
           leadingChildren.add(leadingWidget);
         }
         for (final attached in _attachedFeatures) {
-          leadingChildren.addAll(attached.state._internalBuildLeading());
+          leadingChildren.addAll(attached.state._internalBuildLeading().map(
+                (e) => Focus(
+                  skipTraversal: widget.skipInputFeatureFocusTraversal,
+                  child: e,
+                ),
+              ));
         }
         for (final attached in _attachedFeatures) {
-          trailingChildren.addAll(attached.state._internalBuildTrailing());
+          trailingChildren.addAll(attached.state._internalBuildTrailing().map(
+                (e) => Focus(
+                  skipTraversal: widget.skipInputFeatureFocusTraversal,
+                  child: e,
+                ),
+              ));
         }
         if (trailingWidget != null) {
           trailingChildren.add(trailingWidget);
@@ -1631,7 +1747,7 @@ class TextFieldState extends State<TextField>
     if (widget.onChanged != null) {
       widget.onChanged!(value);
     }
-    formValue = value;
+    formValue = value.isEmpty ? null : value;
     _effectiveText.value = value;
 
     for (final attached in _attachedFeatures) {
@@ -1869,8 +1985,14 @@ class TextFieldState extends State<TextField>
           groupId: widget.groupId,
           onChanged: _onChanged,
           onSelectionChanged: _handleSelectionChanged,
-          onEditingComplete: widget.onEditingComplete,
-          onSubmitted: widget.onSubmitted,
+          onEditingComplete: () {
+            widget.onEditingComplete?.call();
+            _formatSubmit();
+          },
+          onSubmitted: (value) {
+            widget.onSubmitted?.call(value);
+            _formatSubmit();
+          },
           onTapOutside: widget.onTapOutside,
           inputFormatters: formatters,
           rendererIgnoresPointer: true,